refactor(dpp,drive): address 5 review comments — code organization + cleanup
Quantum's review (5 inline comments) on the sum-tree branch HEAD:

1. **index_level/mod.rs: drop the 'NOTE: Copy was dropped'
   commentary** — the trivia about the v3 sum-trees field-add was
   load-bearing during the port but reads as historical noise now
   that the surface is settled.

2. **Extract find_first_countability_change +
   find_first_summability_change to their own file** — both helpers
   are pure pairwise tree-diff routines invoked by validate_update;
   moved to index_level/find_first_change.rs keeping the
   `pub(super)` visibility so validate_update reaches them via the
   sibling module.

3. **Move the 3 e2e test modules out of insert_contract/v0/mod.rs**
   — countable_e2e_tests (606 lines), range_countable_index_e2e_tests
   (3215 lines), and range_summable_index_e2e_tests (257 lines)
   relocate to insert_contract/v0/tests/<name>.rs. v0/mod.rs drops
   from ~4540 lines to ~456, leaving only impl + a single
   `#[cfg(test)] mod tests;` declaration.

4. **Version detect_sum_mode + detect_mode (count side)**: routing
   tables are consensus-relevant on the query surface — protocol
   versions that adjust shape mappings should land behind a
   method-version bump. Added:
   - drive_document_sum_query/mode_detection/{mod.rs,v0/mod.rs} —
     detect_sum_mode dispatcher + v0 impl. Dispatcher reads
     `platform_version.drive.methods.document.query.detect_sum_mode`.
     Old free-fn site in drive_dispatcher.rs gone.
   - drive_document_count_query/mode_detection.rs — added
     `detect_mode_versioned` wrapper that dispatches on
     `detect_count_mode` FeatureVersion to the existing
     `detect_mode` (which becomes the v0 routing table). Existing
     test calls to `detect_mode` stay valid since they pin v0
     directly. count_proof_helpers + count dispatcher switched to
     the versioned wrapper.
   - DriveDocumentQueryMethodVersions grows two FeatureVersion
     fields (detect_count_mode, detect_sum_mode), initialized to 0
     in DRIVE_DOCUMENT_METHOD_VERSIONS_V1/V2/V3.
   AVG has no separate detect function — its dispatcher inlines
   the mode-picking — so no AVG counterpart needed today.

5. **Strip PR 670 references from primary_key_tree_type.rs** — the
   docstring still spoke about "the PR 670 newcomer" and "(per-node
   BOTH; grovedb PR 670)" markers; rephrased to describe the tree
   variants on their own terms now that the PR is merged.

Validation: cargo check --workspace clean; cargo clippy with
-D warnings clean on drive, dpp, drive-proof-verifier; all 3174
drive lib tests + 3471 dpp lib tests pass.
