+++ packages/rs-drive/src/drive/contract/insert/insert_contract/v0/tests/mod.rs
@@ -1,0 +1,21 @@
+//! End-to-end test modules for `insert_contract_operations_v0`.
+//!
+//! Extracted from `v0/mod.rs` (which kept ~4 000 lines of tests
+//! inline alongside ~450 lines of impl) so the impl side stays
+//! readable. Three submodules cover the three feature surfaces
+//! the v0 contract apply path supports today:
+//!
+//! - [`countable_e2e_tests`] — document-type-level
+//!   `documentsCountable` / `rangeCountable` (primary-key tree
+//!   variant).
+//! - [`range_countable_index_e2e_tests`] — per-index
+//!   `rangeCountable: true` (property-name tree variant +
+//!   `NonCounted`-wrapped continuations).
+//! - [`range_summable_index_e2e_tests`] — per-index
+//!   `rangeSummable` / `rangeCountable` 4-way dispatcher
+//!   (`(false, false) | (true, false) | (false, true) | (true,
+//!   true)` corners of the matrix).
+
+mod countable_e2e_tests;
+mod range_countable_index_e2e_tests;
+mod range_summable_index_e2e_tests;
