+++ packages/rs-drive/src/query/drive_document_sum_query/mod.rs
@@ -39,6 +39,9 @@ pub mod drive_dispatcher;
 #[cfg(any(feature = "server", feature = "verify"))]
 pub mod index_picker;
 
+#[cfg(feature = "server")]
+pub mod mode_detection;
+
 #[cfg(any(feature = "server", feature = "verify"))]
 pub mod path_query;
