+++ packages/rs-platform-version/src/version/drive_versions/drive_document_method_versions/v2.rs
@@ -14,6 +14,8 @@ pub const DRIVE_DOCUMENT_METHOD_VERSIONS_V2: DriveDocumentMethodVersions =
             query_contested_documents: 0,
             query_contested_documents_vote_state: 0,
             query_documents_with_flags: 0,
+            detect_count_mode: 0,
+            detect_sum_mode: 0,
         },
         delete: DriveDocumentDeleteMethodVersions {
             add_estimation_costs_for_remove_document_to_primary_storage: 0,
